pyproject.toml
[project]
name = "entsoe-influx"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "entsoe-py>=0.7.1",
    "fastapi>=0.115.0",
    "influxdb-client>=1.35.0",
    "pandas>=2.2.3",
    "python-dateutil>=2.9.0",
    "python-dotenv>=1.2.1",
    "pytz>=2024.1",
    "requests>=2.31.0",
    "typer>=0.12.0",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "ruff>=0.13.3",
]
